pyproject.toml
[project]
name = "mcp101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=2.7.1",
    "fastmcp>=2.10.6",
    "google-genai>=1.25.0",
    "gradio>=5.37.0",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "mcp[cli]>=1.12.0",
    "openai-whisper>=20250625",
    "playsound3>=3.2.4",
    "pyaudio>=0.2.14",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "speechrecognition>=3.14.3",
    "torch>=2.7.1",
    "transformers>=4.53.2",
]
